+++ scripts/scene-packer.js
@@ -144,6 +144,9 @@ export default class ScenePacker {
         adventure: this.adventureName,
       });
 
+      let yesCallback = async () => {
+        await this.importAllContent();
+      };
       if (promptedVersion === '0.0.0') {
         content += game.i18n.format('SCENE-PACKER.welcome.brand-new', {
           adventure: this.adventureName,
@@ -153,6 +156,121 @@ export default class ScenePacker {
           existing: promptedVersion,
           version: moduleVersion,
         });
+        yesCallback = async () => {
+          // The yes callback during an upgrade of a module removes all existing entries that came from this module before
+          // importing everything in again.
+          // Folders are untouched.
+          const scenes = game.scenes.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`));
+          const actors = game.actors.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`));
+          const items = game.items.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`));
+          const journals = game.journal.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`));
+          const rollTables = game.tables.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`));
+          const playlists = game.playlists.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`));
+          const macros = game.macros.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`));
+
+          let listToDelete = '';
+          if (scenes.length) {
+            listToDelete += `<li>${game.i18n.localize('Scenes')}: ${scenes.length}</li>`;
+          }
+          if (actors.length) {
+            listToDelete += `<li>${game.i18n.localize('Actors')}: ${actors.length}</li>`;
+          }
+          if (items.length) {
+            listToDelete += `<li>${game.i18n.localize('Items')}: ${items.length}</li>`;
+          }
+          if (journals.length) {
+            listToDelete += `<li>${game.i18n.localize('Journals')}: ${journals.length}</li>`;
+          }
+          if (rollTables.length) {
+            listToDelete += `<li>${game.i18n.localize('Roll tables')}: ${rollTables.length}</li>`;
+          }
+          if (playlists.length) {
+            listToDelete += `<li>${game.i18n.localize('Playlists')}: ${playlists.length}</li>`;
+          }
+          if (macros.length) {
+            listToDelete += `<li>${game.i18n.localize('Macros')}: ${macros.length}</li>`;
+          }
+          if (listToDelete) {
+            listToDelete = `<p>${game.i18n.localize('SCENE-PACKER.welcome.update-warning-list-to-delete')}</p><ul>${listToDelete}</ul>`;
+          }
+
+          const doReplacement = async function () {
+            const sceneIDs = game.scenes.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`)).map(s => s.id);
+            if (sceneIDs.length) {
+              if (!isNewerVersion('0.8.0', game.data.version)) {
+                await Scene.deleteDocuments(sceneIDs);
+              } else {
+                await Scene.delete(sceneIDs);
+              }
+            }
+            const actorIDs = game.actors.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`)).map(s => s.id);
+            if (actorIDs.length) {
+              if (!isNewerVersion('0.8.0', game.data.version)) {
+                await Actor.deleteDocuments(actorIDs);
+              } else {
+                await Actor.delete(actorIDs);
+              }
+            }
+            const itemIDs = game.items.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`)).map(s => s.id);
+            if (itemIDs.length) {
+              if (!isNewerVersion('0.8.0', game.data.version)) {
+                await Item.deleteDocuments(itemIDs);
+              } else {
+                await Item.delete(itemIDs);
+              }
+            }
+            const journalIDs = game.journal.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`)).map(s => s.id);
+            if (journalIDs.length) {
+              if (!isNewerVersion('0.8.0', game.data.version)) {
+                await JournalEntry.deleteDocuments(journalIDs);
+              } else {
+                await JournalEntry.delete(journalIDs);
+              }
+            }
+            const rollTableIDs = game.tables.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`)).map(s => s.id);
+            if (rollTableIDs.length) {
+              if (!isNewerVersion('0.8.0', game.data.version)) {
+                await RollTable.deleteDocuments(rollTableIDs);
+              } else {
+                await RollTable.delete(rollTableIDs);
+              }
+            }
+            const playlistIDs = game.playlists.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`)).map(s => s.id);
+            if (playlistIDs.length) {
+              if (!isNewerVersion('0.8.0', game.data.version)) {
+                await Playlist.deleteDocuments(playlistIDs);
+              } else {
+                await Playlist.delete(playlistIDs);
+              }
+            }
+            const macroIDs = game.macros.filter(s => s.getFlag('core', 'sourceId')?.startsWith(`Compendium.${this.moduleName}`)).map(s => s.id);
+            if (macroIDs.length) {
+              if (!isNewerVersion('0.8.0', game.data.version)) {
+                await Macro.deleteDocuments(macroIDs);
+              } else {
+                await Macro.delete(macroIDs);
+              }
+            }
+            await this.importAllContent();
+          }.bind(this);
+
+          Dialog.confirm({
+            title: game.i18n.localize('SCENE-PACKER.welcome.update-warning-title'),
+            content: game.i18n.format('SCENE-PACKER.welcome.update-warning-content', {
+              module: moduleName,
+              world: game.world.id,
+              listToDelete: listToDelete,
+            }),
+            yes: doReplacement,
+            no: () => this.logWarn(true, game.i18n.format('SCENE-PACKER.welcome.update-cancelled', {
+              from: promptedVersion,
+              to: moduleVersion,
+              module: moduleName,
+            })),
+            defaultYes: false
+          });
+
+        };
       }
       let label = game.i18n.localize('SCENE-PACKER.welcome.yes-all-fallback');
       if (!isNewerVersion('0.8.0', game.data.version)) {
@@ -170,9 +288,7 @@ export default class ScenePacker {
           yesAll: {
             icon: '<i class="fas fa-check-double"></i>',
             label: label,
-            callback: async () => {
-              await this.importAllContent();
-            },
+            callback: yesCallback,
           },
           choose: {
             icon: '<i class="fas fa-clipboard-check"></i>',