+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v2.3.13
+
+- Changed the way that the initial import dialog shows for module upgrades.
+  - Upgrading an existing Scene Packer packed module will now prompt the end-user to *replace* their imported content rather than importing beside it.
+  - Includes a big scary warning message to do backups first.
+
 ## v2.3.12
 
 - Updated the `Relink compendium entries` macro to handle the case where *only* hyperlink style links exist within a journal.